requirements.txt
fastapi[standard]
psycopg2-binary